+++ rust/aura/src/command/log.rs
@@ -3,19 +3,15 @@
 use crate::command::misc;
 use crate::error::Error;
 use aura_core as core;
+use chrono::NaiveDate;
 use colored::*;
 use i18n_embed::fluent::FluentLanguageLoader;
 use i18n_embed::LanguageLoader;
 use i18n_embed_fl::fl;
+use std::fs::File;
+use std::io::{BufRead, BufReader, BufWriter, Write};
 use std::path::Path;
 use std::process::Command;
-
-/// Open the Pacman/ALPM log in `bat` or `less`.
-pub(crate) fn view(path: &Path) -> Result<(), Error> {
-    let prog = misc::viewer();
-    Command::new(prog).arg(path).status()?;
-    Ok(())
-}
 
 /// Search the Pacman log for a matching string.
 pub(crate) fn search(path: &Path, term: String) -> Result<(), Error> {
@@ -53,6 +49,37 @@ pub(crate) fn info(fll: FluentLanguageLoader, path: &Path, pks: Vec<String>) {
         }
         println!();
     }
+}
+
+/// Output the content of the Pacman/ALPM log, possibly filtered by date.
+pub(crate) fn view(
+    path: &Path,
+    before: Option<NaiveDate>,
+    after: Option<NaiveDate>,
+) -> Result<(), Error> {
+    let file = BufReader::new(File::open(path)?);
+    let mut out = BufWriter::new(std::io::stdout());
+
+    for rline in file.lines() {
+        let line = rline?;
+        if let Ok(date) = line
+            .chars()
+            .skip(1)
+            .take(10)
+            .collect::<String>() // TODO Avoid the collect somehow?
+            .parse::<NaiveDate>()
+        {
+            match (after, before) {
+                (Some(a), Some(b)) if date >= a && date < b => writeln!(out, "{}", line)?,
+                (Some(a), None) if date >= a => writeln!(out, "{}", line)?,
+                (None, Some(b)) if date < b => writeln!(out, "{}", line)?,
+                (None, None) => writeln!(out, "{}", line)?,
+                (_, _) => {}
+            }
+        }
+    }
+
+    Ok(())
 }
 
 fn pad(mult: usize, longest: usize, s: &str) -> usize {